Tighten Services security groups
Get rid of CIDR blocks for Nomad and vm server connections.
We're using security groups everything, so we should be able to limit connections to security groups.

This change creates a new security group for the services box, and then if we enable Nomad (through "var.enable_nomad"),
it adds security group rules for 4647 (Nomad), 8585 (output-processor), 3001 (build-agent) from both Nomad and VM servers